Refactor: CTS: separate ClusterManager class into own source file
... for better code isolation
