+++ cts/lab/Makefile.am
@@ -20,6 +20,7 @@ ctslibdir		= $(pythondir)/cts
 ctslib_PYTHON		= __init__.py		\
 			  CIB.py		\
 			  cib_xml.py		\
+			  ClusterManager.py	\
 			  CM_common.py		\
 			  CM_corosync.py	\
 			  CTS.py		\
